# Conflicts:
#	backend/garbage_collection/views.py
#	backend/picture_building/tests.py
#	backend/student_at_building_on_tour/views.py
#	backend/student_on_tour/scrambled1.png
#	backend/tour/views.py
#	backend/util/data_generators.py

+++ backend/garbage_collection/views.py
@@ -4,8 +4,8 @@ from drf_spectacular.types import OpenApiTypes
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 
-from base.permissions import IsSuperStudent, IsAdmin, ReadOnlyStudent, ReadOnlyOwnerOfBuilding
 from base.models import GarbageCollection, Building
+from base.permissions import IsSuperStudent, IsAdmin, ReadOnlyStudent, ReadOnlyOwnerOfBuilding, OwnerOfBuilding
 from base.serializers import GarbageCollectionSerializer
 from garbage_collection.serializers import GarbageCollectionDuplicateRequestSerializer
 from util.request_response_util import *
@@ -40,7 +40,7 @@ class DefaultGarbageCollection(APIView):
 
 
 class GarbageCollectionIndividualView(APIView):
-    permission_classes = [IsAuthenticated, IsAdmin | IsSuperStudent | ReadOnlyStudent]
+    permission_classes = [IsAuthenticated, IsAdmin | IsSuperStudent | ReadOnlyStudent | ReadOnlyOwnerOfBuilding]
     serializer_class = GarbageCollectionSerializer
 
     @extend_schema(responses=get_docs(GarbageCollectionSerializer))
@@ -51,7 +51,11 @@ class GarbageCollectionIndividualView(APIView):
         garbage_collection_instance = GarbageCollection.objects.filter(id=garbage_collection_id)
         if not garbage_collection_instance:
             return not_found("GarbageCollection")
-        serializer = GarbageCollectionSerializer(garbage_collection_instance[0])
+        garbage_collection_instance = garbage_collection_instance[0]
+
+        self.check_object_permissions(request, garbage_collection_instance.building)
+
+        serializer = GarbageCollectionSerializer(garbage_collection_instance)
         return get_success(serializer)
 
     @extend_schema(responses=delete_docs())
@@ -62,6 +66,7 @@ class GarbageCollectionIndividualView(APIView):
         garbage_collection_instance = GarbageCollection.objects.filter(id=garbage_collection_id)
         if not garbage_collection_instance:
             return not_found("GarbageCollection")
+        self.check_object_permissions(request, garbage_collection_instance[0].building)
         garbage_collection_instance[0].delete()
         return delete_success()
 
@@ -75,6 +80,9 @@ class GarbageCollectionIndividualView(APIView):
             return not_found("GarbageCollection")
 
         garbage_collection_instance = garbage_collection_instance[0]
+
+        self.check_object_permissions(request, garbage_collection_instance.building)
+
         data = request_to_dict(request.data)
 
         set_keys_of_instance(garbage_collection_instance, data, TRANSLATE)
@@ -86,12 +94,6 @@ class GarbageCollectionIndividualView(APIView):
         return patch_success(serializer)
 
 
-@extend_schema(
-    parameters=param_docs({
-        "start-date": ("Filter by start-date", False, OpenApiTypes.DATE),
-        "end-date": ("Filter by end-date", False, OpenApiTypes.DATE),
-    })
-)
 class GarbageCollectionIndividualBuildingView(APIView):
     """
     /building/<buildingid>
@@ -113,6 +115,7 @@ class GarbageCollectionIndividualBuildingView(APIView):
         building_instance = Building.objects.filter(building_id=building_id)
         if not building_instance:
             return not_found("building")
+
         self.check_object_permissions(request, building_instance[0])
 
         filters = {
